requirements.txt
matplotlib==3.2.2
matplotlib-venn==0.11.6
numpy==1.18.5
pandas==1.1.4
pandas-datareader==0.9.0
pandas-gbq==0.13.3
pandas-profiling==1.4.1
Pillow==7.0.0
pytorch-fid==0.2.0
scikit-image==0.16.2
torch==1.7.0+cu101
torchsummary==1.5.1
torchtext==0.3.1
torchvision==0.8.1+cu101
tqdm==4.41.1
progressbar2==3.51.4